pyproject.toml
[tool.poetry]
name = "nobs-canonicalize"
packages = [{include = "nobs_canonicalize"}]
version = "0.5.0"
description = "Canonicalize verbose text strings into clean, deduplicated groups using BERTopic + LLM"
authors = ["Boris Dev <boris.dev@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.11,<3.15"
numpy = "1.26.4"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
pydantic = "^2.7.1"
instructor = "^1.7.0"
openai = "^1.58.1"
diskcache = "^5.6.3"
bertopic = "^0.16.4"
pytest-asyncio = "^0.25.3"
update = "^0.0.1"
more-itertools = "^10.6.0"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.13"
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
